Initial commit
I'll commit a README later. Here's what this is:

1. A solar system simulation, as simple as I can get it
2. A minimalist keyframing interface for animating to Purple Motion's A Touch of Spring

I'll add some attributions to the README, such as for the galaxy provided by Loic Norgeot, which
I've merely compressed a bit, and the solar system sim that I've pilfered with considerable simplification
from Sander Blue's solar-system-threejs project.

The intention is to use this whole thing to spice up a website.